requirements/requirements.txt
Django==3.2.4
django-cors-headers==3.7.0
django-environ==0.4.5
djangorestframework==3.13.1
pyJWT==2.1.0
argon2-cffi==21.1.0
dj-database-url
pillow 
# whitenoise==5.3.0
psycopg2
gunicorn==20.1.0
# djoser
# djangorestframework-simplejwt
